Curator Template
Source: R/memor.R

Curator Template

memor(use_profile = TRUE, logo = NULL, company = NULL,
  short_title = NULL, watermark = NULL, confidential = TRUE,
  logo_height = "1.2cm", watermark_color = "gray",
  footer_on_first_page = TRUE, toc = FALSE, lot = FALSE, lof = FALSE,
  number_sections = TRUE, latex_engine = "xelatex", ...)

Arguments

use_profile
T/F value for whether the user profile in `~/memor-profile.yaml` will be loaded. Default is `TRUE`.

logo
File path to Company logo file. The logo will be printed on the top right corner of the first page and the size will be adjusted to `logo_height`.

company
Company Name as the page footnote. You can input a list of values, such as company name, address, telephone and email.

short_title
A shorten version of the title to be printed on the top-right corner of every page, except the front page.

watermark
Watermark Text to be printed on the page. You can change its color via `watermark_color`.

confidential
T/F value for whether a red confidential sign will be printed.

logo_height
Height of the logo image. This logo image will be scaled to height and the default height is 1.2cm.

watermark_color
Color for the watermark text. Default is "gray".

footer_on_first_page
T/F value for whether the company info footer will be displayed on the front page.

toc
T/F value for table of contents. See ?rmarkdown::pdf_document for details

lot
T/F value for list of tables.

lof
T/F value for list of figures.

number_sections
logical value for wether sections should be numbered. See ?rmarkdown::pdf_document for details

latex_engine
LaTeX engine. See ?rmarkdown::pdf_document for details

...
other options to be passed to rmarkdown::pdf_document. See ?rmarkdown::pdf_document for details