Memor PDF Memo Template

Memor PDF Memo Template

pdf_memo(use_profile = TRUE, logo = NULL, company = NULL,
  short_title = NULL, watermark = NULL, confidential = FALSE,
  libertine = FALSE, chinese = FALSE, logo_height = "1.2cm",
  watermark_color = "gray", footer_on_first_page = TRUE, toc = FALSE,
  lot = FALSE, lof = FALSE, fancy_captions = TRUE,
  number_sections = TRUE, latex_engine = "xelatex", ...)

Arguments

use_profile
T/F value for whether the user profile in ~/memor-profile.yaml will be loaded. Default is TRUE.

logo
File path to Company logo file. The logo will be printed on the top right corner of the first page and the size will be adjusted to logo_height.

company
Company Name as the page footnote. You can input a list of values, such as company name, address, telephone and email.

short_title
A shorten version of the title to be printed on the top-right corner of every page, except the front page.

watermark
Watermark Text to be printed on the page. You can change its color via watermark_color.

confidential
T/F value for whether a red confidential sign will be printed.

libertine
T/F. Libertine is a collection of open fonts for western languages. We found it fits this template quite well. Default is FALSE though we highly recommend you to turn it on.

chinese
T/F for Chinese language support. Default is FALSE.

logo_height
Height of the logo image. This logo image will be scaled to height and the default height is 1.2cm.

watermark_color
Color for the watermark text. Default is "gray".

footer_on_first_page
T/F value for whether the company info footer will be displayed on the front page.

toc
T/F value for table of contents. See ?rmarkdown::pdf_document for details

lot
T/F value for list of tables.

lof
T/F value for list of figures.

fancy_captions
T/F value for whether to apply special formatting to captions.

number_sections
T/F value for whether sections should be numbered. See ?rmarkdown::pdf_document for details

latex_engine
LaTeX engine. See ?rmarkdown::pdf_document for details

...
other options to be passed to rmarkdown::pdf_document. See ?rmarkdown::pdf_document for details

Examples

if (FALSE) {
library(rmarkdown)
 example <- system.file("example/demo.Rmd", package = "memor")
 render(example, memor::pdf_memo())
}